+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Loading spinner to the Categories component
+
 ## [0.8.0] - 2020-06-09
 
 ### Added